+++ src/config.lua
@@ -1,6 +1,6 @@
 ---@meta zannc-config-Practical_Gods
 return {
-	-- ArtemisAttackCritScale = true,
-	-- ArtemisSpecialCritScale = true,
 	enabled = true,
+}, {
+	enabled = "Disable/Enable the mod, other config options delegated to parent mod: Droppable_Gods",
 }